# Improve single element indexing (and other quality of life improvements)