--- a/src/indexing.jl
+++ b/src/indexing.jl
@@ -88,6 +88,7 @@ is_canonical(::Type{Int}) = static(true)
 is_canonical(::Type{<:StaticInt}) = static(true)
 is_canonical(::Type{StepRange{Int,Int}}) = static(true)
 is_canonical(::Type{UnitRange{Int}}) = static(true)
+is_canonical(::Type{OneTo{Int}}) = static(true)
 is_canonical(::Type{<:OptionallyStaticRange}) = static(true)
 is_canonical(::Type{Vector{Int}}) = static(true)
 is_canonical(::Type{<:Slice}) = static(true)
@@ -95,6 +96,25 @@ is_canonical(::Type{<:Slice}) = static(true)
 _is_canon(::Type{T}, ::StaticInt{0}) where {T} = static(true)
 @inline function _is_canon(::Type{T}, n::StaticInt{N}) where {T,N}
     return is_canonical(_get_tuple(T, n)) & _is_canon(T, n - static(1))
+end
+
+"""
+    canonicalize(x)
+
+Checks if `x` is in a canonical form for indexing. If `x` is already in a canonical form
+then it is returned unchanged. If `x` is not in a canonical form then it is passed to
+`canonical_convert`.
+"""
+canonicalize(x) = _canonicalize(is_canonical(x), x)
+_canonicalize(::True, x) = x
+_canonicalize(::False, x) = canonical_convert(x)
+
+canonical_convert(x::Integer) = convert(Int, x)
+function canonical_convert(x::AbstractRange)
+    return OptionallyStaticStepRange(static_first(x), static_step(x), static_last(x))
+end
+function canonicalize_convert(x::AbstractUnitRange{<:Integer})
+    return OptionallyStaticUnitRange(static_first(x), static_last(x))
 end
 
 _layout(::IndexLinear, x::Tuple) = LinearIndices(x)
@@ -288,7 +308,7 @@ end
 end
 @propagate_inbounds function to_index(::IndexLinear, x, arg::Integer)
     @boundscheck checkindex(Bool, indices(x), arg) || throw(BoundsError(x, arg))
-    return _int(arg)
+    return canonicalize(arg)
 end
 @propagate_inbounds function to_index(::IndexLinear, axis, arg::AbstractArray{Bool})
     @boundscheck checkbounds(axis, arg)
@@ -336,9 +356,9 @@ to_index(::IndexCartesian, x, i::Integer) = NDIndex(_int2subs(offsets(x), size(x
 @inline function _int2subs(o::Tuple{Any,Vararg{Any}}, s::Tuple{Any,Vararg{Any}}, i)
     len = first(s)
     inext = div(i, len)
-    return (_int(i - len * inext + first(o)), _int2subs(tail(o), tail(s), inext)...)
+    return (canonicalize(i - len * inext + first(o)), _int2subs(tail(o), tail(s), inext)...)
 end
-_int2subs(o::Tuple{Any}, s::Tuple{Any}, i) = _int(i + first(o))
+_int2subs(o::Tuple{Any}, s::Tuple{Any}, i) = canonicalize(i + first(o))
 
 """
     unsafe_reconstruct(A, data; kwargs...)