While working on layouts for #141 I realized that unless `getindex` was skipped (e.g., directly calling `unsafe_getindex`) than element-wise indexing was a lot slower. I approached this from a lot of different angles (some of which will probably come up in another PR still), but the issue appears to be due to the compiler not knowing to elide reconstruction of something like `Tuple{Int,Int}`, despite not undergoing any change throughout `to_indices`. So I'll start with the fix I had for that.

# Canonical Indexing Types
I chose the name because I saw a similar concept in [LoopVectorization](https://github.com/JuliaSIMD/LoopVectorization.jl/blob/2fdb50f8685c85f95e3292b137c0a8f3a9ca6c48/src/modeling/graphs.jl#L873). In `to_indices(A, args::Tuple)` each element of `args` is checked to see if it is a canonical indexing type up front, avoiding recursive reconstruction of the tuple of indexers. This check is performed by `is_canonical`, resulting in the following improvement:

```julia
x = ones(4,4,4);

#  current version
julia> @btime ArrayInterface.getindex($x, 3, 3, 3)
  4.566 ns (0 allocations: 0 bytes)
1.0

#  new version
julia> @btime ArrayInterface.getindex($x, 3, 3, 3)
  1.755 ns (0 allocations: 0 bytes)
1.0

# base version
julia> @btime getindex($x, 3, 3, 3)
  1.756 ns (0 allocations: 0 bytes)
1.0

```

I've also expanded on the number of types that are considered "canonical" indexers from that of base (e.g., we don't ever need to change `OptionallyStaticUnitRange`), so there may also be some (_very small_) improvements on indexing collection.
I think we could still squeeze some more performance out of this by avoiding construction of axes for types where the conversion to the canonical state is known already.
This is kind of the idea behind `canonicalize(x)`, which converts `x` to a canonical indexer and helped clean up some internal code that does the same thing.
I'm probably going to work on that optimization in another PR b/c it will probably require some involved work on the internal mechanics of `to_indices`.
For now we at least are as fast as Base.

# LazyAxis

A lot of array traits can be derived from looking an array's axes.
Many arrays don't have fully constructed axes stored with them, meaning we usually spend some unnecessary time composing an axis to just find out its size or offset.
Using `lazy_axes` we can sometimes avoid these costs.

For example, we can avoid the cost of retrieving size information for an `Array`.

```julia
axes_map_first(x) = map(first, ArrayInterface.axes(x))
lazy_axes_map_first(x) = map(first, ArrayInterface.lazy_axes(x))

julia> @btime axes_map_first($x)
  1.754 ns (0 allocations: 0 bytes)
(1, 1, 1)

julia> @btime lazy_axes_map_first($x)
  0.046 ns (0 allocations: 0 bytes)
(1, 1, 1)
```

Much of this improvement is lost when we size information has to be called by both approaches.
```julia
axes_map_last(x) = map(last, ArrayInterface.axes(x))
lazy_axes_map_last(x) = map(last, ArrayInterface.lazy_axes(x))

julia> @btime axes_map_last($x)
  1.608 ns (0 allocations: 0 bytes)
(4, 4, 4)

julia> @btime lazy_axes_map_last($x)
  1.470 ns (0 allocations: 0 bytes)
(4, 4, 4)
```

I've started putting this in several places and it looks like it either does as well as the eager method or improves on it.
I could probably be a little more exhaustive with improving the performance of this, but figured it was helpful enough to include as is right now.

One final note, I'm aware that I've added a bunch of messy code for bounds checking. I'm going to clean this up when I fine tune `to_indices` later on.