@chriselrod 2021-05-03
As an unrelated comment, I noticed this recently:
```julia
julia> using ArrayInterface

julia> ci = CartesianIndices(ntuple(i -> Base.OneTo(32), 4));

julia> cis = CartesianIndices(ntuple(i -> ArrayInterface.One():32, 4));

julia> @btime sum($ci)
  827.116 μs (0 allocations: 0 bytes)
CartesianIndex(17301504, 17301504, 17301504, 17301504)

julia> @btime sum($cis)
  915.828 ms (8655868 allocations: 376.07 MiB)
CartesianIndex(17301504, 17301504, 17301504, 17301504)
```
This happens when iterating over `CartesianIndices` with `OptionallyStaticUnitRange{StaticInt{F}}`. For now, I've switched `StrideArraysCore` to use `Base.OneTo` for 1-indexed axes.
The issue is a type instability caused by a type instability in [__inc](https://github.com/JuliaLang/julia/blob/d22c3041a72e62ddc3b9e17e83a61b841a6d413b/base/multidimensional.jl#L406-L416):
```julia
    @inline function __inc(state::Tuple{Int,Int,Vararg{Int,N}}, indices::Tuple{OrdinalRangeInt,OrdinalRangeInt,Vararg{OrdinalRangeInt,N}}) where {N}
        rng = indices[1]
        I = state[1] + step(rng)
        if __is_valid_range(I, rng) && state[1] != last(rng)
            return true, (I, tail(state)...)
        end
        t1, t2 = tail(state), tail(indices)
        # avoid dynamic dispatch by telling the compiler relational invariants
        valid, I = isa(t1, Tuple{Int}) ? __inc(t1, t2::Tuple{OrdinalRangeInt}) : __inc(t1, t2::Tuple{OrdinalRangeInt,OrdinalRangeInt,Vararg{OrdinalRangeInt}})
        return valid, (first(rng), I...)
    end
```
`state[1] + step(rng)` and `first(rng)` won't have the same type.
Could be fixed by making `first(::OptionallyStaticUnitRange)` always return `Int`.