@chriselrod 2021-05-03
> I approached this from a lot of different angles (some of which will probably come up in another PR still), but the issue appears to be due to the compiler not knowing to elide reconstruction of something like `Tuple{Int,Int}`, despite not undergoing any change throughout to_indices. So I'll start with the fix I had for that.

Do you have an example?
All of this should inline, in which case it should be able to decompose all the `Tuple`s into individual scalar operations.

> I chose the name because I saw a similar concept in [LoopVectorization](https://github.com/JuliaSIMD/LoopVectorization.jl/blob/2fdb50f8685c85f95e3292b137c0a8f3a9ca6c48/src/modeling/graphs.jl#L873)

It should largely be unnecessary, because the `known_*` functions should cover all the cases when it comes time for LoopVectorization to [process the loops](https://github.com/JuliaSIMD/LoopVectorization.jl/blob/3f605b94acb9c700f17166b8867316661e02082d/src/reconstruct_loopset.jl#L35-L54).
The fact that there are different methods there is somewhat a hold over from before it began using the `ArrayInterface.known_*` methods.

> A lot of array traits can be derived from looking an array's axes.
Many arrays don't have fully constructed axes stored with them, meaning we usually spend some unnecessary time composing an axis to just find out its size or offset.

This seems like it really ought to be compiled away.
Seems like a benchmarking artifact:
```julia
julia> = rand(4,5);

julia> @inline sum_first_axes(A) = sum(map(first, ArrayInterface.axes(A)))
sum_first_axes (generic function with 1 method)

julia> @btime sum_first_axes($A)
  1.410 ns (0 allocations: 0 bytes)
2

julia> @code_typed sum_first_axes(A)
CodeInfo(
1 ─     Base.arraysize(A, 1)::Int64
│       Base.arraysize(A, 2)::Int64
└──     return 2
) => Int64

julia> @code_llvm sum_first_axes(A)
;  @ REPL[23]:1 within `sum_first_axes'
define i64 @julia_sum_first_axes_1792({}* nonnull align 16 dereferenceable(40) %0) #0 {
top:
  ret i64 2
}

julia> @code_native sum_first_axes(A)
        .text
; ┌ @ REPL[23]:1 within `sum_first_axes'
        movl    $2, %eax
        retq
        nopw    %cs:(%rax,%rax)
; └
```
1 clock cycle is around 0.25ns, so anything less than that means the benchmark was completely compiled away.
Seems like that doesn't happen with the current axes in a benchmark, but that otherwise everything does disappear.