@chriselrod 2021-05-03
I still need to spend the time to look over your other PR, but this one looks fine to me.
I'm generally in favor of being easier on the compiler. If a benchmark doesn't get compiled away with `axes` but does with `lazy_axes`, there're probably more cases where the compiler ends up doing better with the latter.