@Tokazama 2021-05-03
> Do you have an example?
All of this should inline, in which case it should be able to decompose all the Tuples into individual scalar operations.

I always operated under that assumption until I started benchmarking everything.

```julia
# old
@btime ArrayInterface.to_indices($x, (3,3,3))
julia> @btime ArrayInterface.to_indices($x, (3,3,3))
  4.453 ns (0 allocations: 0 bytes)

# new
julia> @btime ArrayInterface.to_indices($x, (3,3,3))
  2.045 ns (0 allocations: 0 bytes)
(3, 3, 3)
```

Apparently it's a [known issue](https://github.com/JuliaLang/julia/blob/b3b9aa0eb1a90765dd3d395073992c534254a23f/base/multidimensional.jl#L799)