@Tokazama 2021-05-03
> I still need to spend the time to look over your other PR

Just get to that when you have a chance. I recently discovered the [Data Layout Modeling](https://mlir.llvm.org/docs/DataLayout/) section of the MLIR manual and I I'm considering changing a couple of things in how I approached that.